Rename sigmoid to sigmoid_tools
