+++ simple_neuralnet_python/__init__.py
@@ -1,4 +1,4 @@
 from simple_neuralnet_python.neuralnet import NeuralNet
-from simple_neuralnet_python.sigmoid import sigmoid, sigmoid_derivative
+from simple_neuralnet_python.sigmoid_tools import sigmoid, sigmoid_derivative
 
 __all__ = ['NeuralNet', 'sigmoid', 'sigmoid_derivative']
