+++ simple_neuralnet_python/sigmoid_tools.py
